chore(CommonDomain): update tick duration

+++ common-domain/src/main/kotlin/ru/woowy/game/GameSpeed.kt
@@ -1,7 +1,7 @@
 package ru.woowy.game
 
 enum class GameSpeed(
-    val gameMinutesPerTick: Int,
+    val gameMinutesPerTick: Long,
 ) {
     REAL_TIME(10),
     ACCELERATED(60),
